requirements.txt
streamlit==1.54.0
openai==2.20.0
requests==2.32.5
python-dotenv==1.2.1
numpy
faiss-cpu
sentence-transformers
torch